Make multiple operands work
This took a long time, without tests I'm not sure how I would have made this work at all.
The code is not perfect, but it was the simplest thing I could do to make it work. I'm not even sure
I'll refactor at this point, I can't immediately see a better way to express things.  But future
tests might make it clearer to me.