Factor away the flag for resetting the display
I noticed that the expectingOperand flag and resetDisplay flags were set and cleared in the same
places, which gave me a clue that I only needed one of them.  Because I had tests, I could try that
theory easily and quickly. In doing so, I found that I had to set the initial value of expectingOperand
to false, but again, with tests I could very in under a second that this was a safe change to make.